+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # RealmJoin Remediation Scripts Changelog
 
+## 2025-04-16
+- Added Remove HP applications
+- Added Verify and trigger Windows activation
+
 ## 2025-02-14
-- Improved: Deactivate the management of M365 Apps through cloud update
+- Improved Deactivate the management of M365 Apps through cloud update
 - Removed as not needed: Activate the management of M365 Apps through cloud update
 
 ## 2025-02-11